backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv
sqlalchemy>=2.0
asyncpg
alembic
redis
passlib[argon2]
aiokafka
sqlalchemy-utils
google-generativeai
python-multipart
minio